feat: remove declarations
